The cycles task1, task2,task4,task6 is recovered. In the future i will write the name of the class correct.

+++ com.epam.learn.jbupskill.lab/src/module1/cycles/task4.java
@@ -6,8 +6,8 @@ import java.math.*;
 public class task4 {
     public static void main(String[] args) {
         BigInteger sumProduct = BigInteger.valueOf(1);
-        for (int i = 1; i <= 200; i++) {
-            sumProduct = sumProduct.multiply(BigInteger.valueOf((long) i * i));
+        for (long i = 1; i <= 200; i++) {
+            sumProduct = sumProduct.multiply(BigInteger.valueOf(i * i));
         }
         System.out.println("Sum product of number = " + sumProduct);
     }
